backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
SQLAlchemy==2.0.35
alembic==1.13.2
PyMySQL==1.1.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
celery[redis]==5.4.0
redis==5.0.8
pyrogram==2.0.106
tgcrypto==1.2.5
faker==29.0.0
prometheus-client==0.20.0
slowapi==0.1.9
sentry-sdk[fastapi]==2.17.0
stripe==10.9.0
